refactor(cart): change name for path(NO-JIRA)
