+++ src/app/app.component.spec.ts
@@ -3,7 +3,7 @@ import { provideMockStore } from '@ngrx/store/testing'
 
 import { AppComponent } from './app.component'
 import { AuthFacade } from './auth/auth-store/auth.facade'
-import { CartFacade } from './cart/cart-store copy/cart.facade'
+import { CartFacade } from './cart/cart-store/cart.facade'
 
 describe('AppComponent', () => {
   beforeEach(async () => {
