+++ src/app/app.module.ts
@@ -12,9 +12,9 @@ import { AppComponent } from './app.component'
 import { AuthEffects } from './auth/auth-store/auth.effects'
 import { AuthFacade } from './auth/auth-store/auth.facade'
 import { authReducer } from './auth/auth-store/auth.reducer'
-import { CartEffects } from './cart/cart-store copy/cart-store.effects'
-import { cartReducer } from './cart/cart-store copy/cart-store.reducer'
-import { CartFacade } from './cart/cart-store copy/cart.facade'
+import { CartEffects } from './cart/cart-store/cart-store.effects'
+import { cartReducer } from './cart/cart-store/cart-store.reducer'
+import { CartFacade } from './cart/cart-store/cart.facade'
 import { HeaderComponent } from './core/components/header/header.component'
 import { CoreModule } from './core/core.module'
 import { StorageModule } from './core/storage/storage.module'
